requirements.txt
PyQt6>=6.4.0
pyserial>=3.5
numpy>=1.21.0
scipy>=1.7.0
mujoco>=2.3.0
